Pulled buildImgTemplate function into utils file
